+++ packages/packages/apps/onboarding/src/steps/screens/__tests__/building-for.test.tsx
@@ -1,138 +1,15 @@
 /* eslint-disable testing-library/no-test-id-queries */
-import * as React from 'react';
-import { __deleteStore } from '@elementor/store';
-import { QueryClient } from '@tanstack/react-query';
-import { fireEvent, render, screen, waitFor } from '@testing-library/react';
+import { fireEvent, screen, waitFor } from '@testing-library/react';
 
-import { App } from '../../../components/app';
-
-jest.mock( '@elementor/query', () => {
-	const actual = jest.requireActual( '@elementor/query' );
-	return {
-		...actual,
-		createQueryClient: () =>
-			new QueryClient( {
-				defaultOptions: {
-					queries: {
-						refetchOnWindowFocus: false,
-						refetchOnReconnect: false,
-					},
-				},
-			} ),
-	};
-} );
-
-// Mock fetch for API calls
-const mockFetch = jest.fn();
-global.fetch = mockFetch;
-
-interface OnboardingConfig {
-	version: string;
-	restUrl: string;
-	nonce: string;
-	steps: Array< {
-		id: string;
-		label: string;
-		type: 'single' | 'multiple';
-	} >;
-	progress: {
-		current_step_id?: string;
-		current_step_index?: number;
-		completed_steps?: string[];
-	};
-	choices: Record< string, unknown >;
-	hadUnexpectedExit: boolean;
-	isConnected: boolean;
-	userName?: string;
-	urls: {
-		dashboard: string;
-		editor: string;
-		connect: string;
-	};
-}
-
-const defaultConfig: OnboardingConfig = {
-	version: '1.0.0',
-	restUrl: 'https://test.local/wp-json/elementor/v1/e-onboarding/',
-	nonce: 'test-nonce',
-	steps: [
-		{
-			id: 'building_for',
-			label: 'Who are you building for?',
-			type: 'single',
-		},
-		{
-			id: 'site_about',
-			label: 'What is your site about?',
-			type: 'multiple',
-		},
-		{
-			id: 'experience_level',
-			label: 'Experience level',
-			type: 'single',
-		},
-		{
-			id: 'theme_selection',
-			label: 'Theme selection',
-			type: 'single',
-		},
-		{
-			id: 'site_features',
-			label: 'Site features',
-			type: 'multiple',
-		},
-	],
-	progress: {
-		current_step_id: 'building_for',
-		current_step_index: 0,
-		completed_steps: [],
-	},
-	choices: {},
-	hadUnexpectedExit: false,
-	isConnected: true,
-	urls: {
-		dashboard: 'https://test.local/wp-admin/',
-		editor: 'https://test.local/editor',
-		connect: 'https://test.local/connect',
-	},
-};
-
-type ConfigOverrides = Partial< OnboardingConfig >;
-
-const createMockConfig = ( overrides: ConfigOverrides = {} ): { 'e-onboarding': OnboardingConfig } => ( {
-	'e-onboarding': {
-		...defaultConfig,
-		...overrides,
-		progress: {
-			...defaultConfig.progress,
-			...overrides.progress,
-		},
-	},
-} );
+import { mockFetch, renderApp, setupOnboardingTests } from '../../../__tests__/test-utils';
 
 describe( 'BuildingFor', () => {
-	beforeEach( () => {
-		__deleteStore();
-		mockFetch.mockReset();
-		mockFetch.mockResolvedValue( {
-			ok: true,
-			json: () => Promise.resolve( { success: true } ),
-		} );
-	} );
-
-	afterEach( () => {
-		window.elementorAppConfig = undefined;
-		__deleteStore();
-	} );
+	setupOnboardingTests();
 
 	describe( 'Greeting text', () => {
 		it( 'should show greeting without name for guest user', () => {
 			// Arrange
-			window.elementorAppConfig = createMockConfig( {
-				isConnected: false,
-			} );
-
-			render( <App /> );
+			renderApp( { isConnected: false } );
 
 			// Act - continue as guest to reach the building_for step
 			fireEvent.click( screen.getByText( 'Continue as a guest' ) );
@@ -143,14 +20,8 @@ describe( 'BuildingFor', () => {
 		} );
 
 		it( 'should show greeting without name when connected user has no userName', () => {
-			// Arrange
-			window.elementorAppConfig = createMockConfig( {
-				isConnected: true,
-				userName: '',
-			} );
-
-			// Act
-			render( <App /> );
+			// Arrange & Act
+			renderApp( { isConnected: true, userName: '' } );
 
 			// Assert
 			expect( screen.getByTestId( 'building-for-step' ) ).toBeInTheDocument();
@@ -158,14 +29,8 @@ describe( 'BuildingFor', () => {
 		} );
 
 		it( 'should show greeting with name when connected user has userName', () => {
-			// Arrange
-			window.elementorAppConfig = createMockConfig( {
-				isConnected: true,
-				userName: 'John',
-			} );
-
-			// Act
-			render( <App /> );
+			// Arrange & Act
+			renderApp( { isConnected: true, userName: 'John' } );
 
 			// Assert
 			expect( screen.getByTestId( 'building-for-step' ) ).toBeInTheDocument();
@@ -175,13 +40,8 @@ describe( 'BuildingFor', () => {
 
 	describe( 'Option rendering', () => {
 		it( 'should render all building_for options', () => {
-			// Arrange
-			window.elementorAppConfig = createMockConfig( {
-				isConnected: true,
-			} );
-
-			// Act
-			render( <App /> );
+			// Arrange & Act
+			renderApp( { isConnected: true } );
 
 			// Assert
 			expect( screen.getByText( 'Myself or someone I know' ) ).toBeInTheDocument();
@@ -191,13 +51,8 @@ describe( 'BuildingFor', () => {
 		} );
 
 		it( 'should render the step title', () => {
-			// Arrange
-			window.elementorAppConfig = createMockConfig( {
-				isConnected: true,
-			} );
-
-			// Act
-			render( <App /> );
+			// Arrange & Act
+			renderApp( { isConnected: true } );
 
 			// Assert
 			expect( screen.getByText( 'Who are you building for?' ) ).toBeInTheDocument();
@@ -207,11 +62,7 @@ describe( 'BuildingFor', () => {
 	describe( 'Selection behavior', () => {
 		it( 'should call API with correct value when selecting an option', async () => {
 			// Arrange
-			window.elementorAppConfig = createMockConfig( {
-				isConnected: true,
-			} );
-
-			render( <App /> );
+			renderApp( { isConnected: true } );
 
 			// Act
 			fireEvent.click( screen.getByText( 'Myself or someone I know' ) );
@@ -230,11 +81,7 @@ describe( 'BuildingFor', () => {
 
 		it( 'should navigate to next step after successful selection', async () => {
 			// Arrange
-			window.elementorAppConfig = createMockConfig( {
-				isConnected: true,
-			} );
-
-			render( <App /> );
+			renderApp( { isConnected: true } );
 
 			// Act
 			fireEvent.click( screen.getByText( 'A client' ) );
@@ -253,11 +100,7 @@ describe( 'BuildingFor', () => {
 
 		it( 'should call API with correct value for different options', async () => {
 			// Arrange
-			window.elementorAppConfig = createMockConfig( {
-				isConnected: true,
-			} );
-
-			render( <App /> );
+			renderApp( { isConnected: true } );
 
 			// Act
 			fireEvent.click( screen.getByText( 'Just exploring' ) );
@@ -277,14 +120,11 @@ describe( 'BuildingFor', () => {
 
 	describe( 'Pre-selected state', () => {
 		it( 'should show previously selected option from saved choices', () => {
-			// Arrange
-			window.elementorAppConfig = createMockConfig( {
+			// Arrange & Act
+			renderApp( {
 				isConnected: true,
 				choices: { building_for: 'business' },
 			} );
-
-			// Act
-			render( <App /> );
 
 			// Assert - the business option should be pressed
 			const businessButton = screen.getByRole( 'button', { name: 'My business or workplace' } );
